services/worker/requirements.txt
# Celery Worker Requirements

# Celery & Broker
celery==5.3.6
redis==5.0.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# PDF Processing
pymupdf==1.23.8

# OCR
pytesseract==0.3.10
pillow==10.2.0

# Object Storage
boto3==1.34.0
minio==7.2.0

# Validation & Parsing
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
python-dateutil==2.8.2

# Utils
python-magic==0.4.27
requests==2.31.0

# Logging & Monitoring
structlog==24.1.0
prometheus-client==0.19.0

# Development
pytest==7.4.4
pytest-celery==0.0.0
black==24.1.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Security
cryptography==42.0.0
